Add asAsset to CaptureContent Extract enums to separate file
